# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Doula {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  location: string;
  bio: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  videoIntroUrl?: string | null;
  serviceTypes: string[];
  /** @nullable */
  yearsExperience?: number | null;
  /** @nullable */
  birthsAttended?: number | null;
  certifications?: string[];
  trainings?: string[];
  specialties?: string[];
  languages?: string[];
  /** @nullable */
  rateMin?: number | null;
  /** @nullable */
  rateMax?: number | null;
  acceptingClients: boolean;
  insuranceAccepted?: boolean;
  slidingScaleAvailable?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  profileCompleteness: number;
  /** @nullable */
  philosophyStatement?: string | null;
  /** @nullable */
  approachDescription?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
  featured?: boolean;
  createdAt: string;
}

export interface DoulaInput {
  name: string;
  email: string;
  phone?: string;
  location: string;
  bio: string;
  tagline?: string;
  photoUrl?: string;
  videoIntroUrl?: string;
  serviceTypes: string[];
  yearsExperience?: number;
  birthsAttended?: number;
  certifications?: string[];
  trainings?: string[];
  specialties?: string[];
  languages?: string[];
  rateMin?: number;
  rateMax?: number;
  acceptingClients?: boolean;
  insuranceAccepted?: boolean;
  slidingScaleAvailable?: boolean;
  philosophyStatement?: string;
  approachDescription?: string;
  website?: string;
  instagramHandle?: string;
}

export interface DoulaUpdate {
  name?: string;
  phone?: string;
  location?: string;
  bio?: string;
  tagline?: string;
  photoUrl?: string;
  videoIntroUrl?: string;
  serviceTypes?: string[];
  yearsExperience?: number;
  birthsAttended?: number;
  certifications?: string[];
  trainings?: string[];
  specialties?: string[];
  languages?: string[];
  rateMin?: number;
  rateMax?: number;
  acceptingClients?: boolean;
  insuranceAccepted?: boolean;
  slidingScaleAvailable?: boolean;
  philosophyStatement?: string;
  approachDescription?: string;
  website?: string;
  instagramHandle?: string;
  featured?: boolean;
}

export type DoulaServiceServiceType =
  (typeof DoulaServiceServiceType)[keyof typeof DoulaServiceServiceType];

export const DoulaServiceServiceType = {
  birth: "birth",
  postpartum: "postpartum",
  education: "education",
  other: "other",
} as const;

/**
 * @nullable
 */
export type DoulaServiceRateType =
  | (typeof DoulaServiceRateType)[keyof typeof DoulaServiceRateType]
  | null;

export const DoulaServiceRateType = {
  flat: "flat",
  hourly: "hourly",
  package: "package",
} as const;

export interface DoulaService {
  id: number;
  doulaId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  serviceType: DoulaServiceServiceType;
  /** @nullable */
  rateType?: DoulaServiceRateType;
  /** @nullable */
  rate?: number | null;
  /** @nullable */
  includedHours?: number | null;
  /** @nullable */
  packageDetails?: string | null;
}

export type DoulaServiceInputServiceType =
  (typeof DoulaServiceInputServiceType)[keyof typeof DoulaServiceInputServiceType];

export const DoulaServiceInputServiceType = {
  birth: "birth",
  postpartum: "postpartum",
  education: "education",
  other: "other",
} as const;

export type DoulaServiceInputRateType =
  (typeof DoulaServiceInputRateType)[keyof typeof DoulaServiceInputRateType];

export const DoulaServiceInputRateType = {
  flat: "flat",
  hourly: "hourly",
  package: "package",
} as const;

export interface DoulaServiceInput {
  name: string;
  description?: string;
  serviceType: DoulaServiceInputServiceType;
  rateType?: DoulaServiceInputRateType;
  rate?: number;
  includedHours?: number;
  packageDetails?: string;
}

export interface Review {
  id: number;
  doulaId: number;
  clientName: string;
  rating: number;
  /** @nullable */
  title?: string | null;
  body: string;
  serviceType: string;
  /** @nullable */
  birthYear?: number | null;
  verified?: boolean;
  createdAt: string;
}

export interface ReviewInput {
  clientName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  title?: string;
  body: string;
  serviceType: string;
  birthYear?: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  doulaId: number;
  /** @nullable */
  doulaName?: string | null;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  clientPhone?: string | null;
  serviceType: string;
  /** @nullable */
  dueDate?: string | null;
  status: BookingStatus;
  message: string;
  /** @nullable */
  doulaResponse?: string | null;
  createdAt: string;
}

export interface BookingInput {
  doulaId: number;
  clientName: string;
  clientEmail: string;
  clientPhone?: string;
  serviceType: string;
  dueDate?: string;
  message: string;
}

export type BookingUpdateStatus =
  (typeof BookingUpdateStatus)[keyof typeof BookingUpdateStatus];

export const BookingUpdateStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  doulaResponse?: string;
}

export interface PlatformStats {
  totalDoulas: number;
  acceptingClients: number;
  averageRating: number;
  totalReviews: number;
  totalBookings: number;
  birthDoulaCount?: number;
  postpartumDoulaCount?: number;
}

export interface DoulaDashboard {
  doulaId: number;
  pendingBookings: number;
  acceptedBookings: number;
  completedBookings: number;
  totalReviews: number;
  /** @nullable */
  averageRating: number | null;
  profileCompleteness: number;
  recentBookings: Booking[];
  recentReviews: Review[];
}

export type ListDoulasParams = {
  serviceType?: ListDoulasServiceType;
  language?: string;
  minRating?: number;
  maxRate?: number;
  acceptingClients?: boolean;
  search?: string;
};

export type ListDoulasServiceType =
  (typeof ListDoulasServiceType)[keyof typeof ListDoulasServiceType];

export const ListDoulasServiceType = {
  birth: "birth",
  postpartum: "postpartum",
  birth_and_postpartum: "birth_and_postpartum",
} as const;

export type ListBookingsParams = {
  doulaId?: number;
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  completed: "completed",
  cancelled: "cancelled",
} as const;