# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BookYourDoula API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doulas/{id}/dashboard:
    get:
      operationId: getDoulasDashboard
      tags: [discovery]
      summary: Doula dashboard summary (bookings, reviews, profile completeness)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoulaDashboard"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List booking requests
      parameters:
        - in: query
          name: doulaId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, declined, completed, cancelled]
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status or details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
components:
  schemas:
...
    Review:
      type: object
      required:
        - id
        - doulaId
        - clientName
        - rating
        - body
        - serviceType
        - createdAt
      properties:
        id:
          type: integer
        doulaId:
          type: integer
        clientName:
          type: string
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: string
        serviceType:
          type: string
        birthYear:
          type: ["integer", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
...
    Booking:
      type: object
      required:
        - id
        - doulaId
        - clientName
        - clientEmail
        - serviceType
        - status
        - message
        - createdAt
      properties:
        id:
          type: integer
        doulaId:
          type: integer
        doulaName:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: ["string", "null"]
        serviceType:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined, completed, cancelled]
        message:
          type: string
        doulaResponse:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required:
        - doulaId
        - clientName
        - clientEmail
        - serviceType
        - message
      properties:
        doulaId:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        serviceType:
          type: string
        dueDate:
          type: string
        message:
          type: string
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, declined, completed, cancelled]
        doulaResponse:
          type: string
...
    DoulaDashboard:
      type: object
      required:
        - doulaId
        - pendingBookings
        - acceptedBookings
        - completedBookings
        - totalReviews
        - averageRating
        - profileCompleteness
        - recentBookings
        - recentReviews
      properties:
        doulaId:
          type: integer
        pendingBookings:
          type: integer
        acceptedBookings:
          type: integer
        completedBookings:
          type: integer
        totalReviews:
          type: integer
        averageRating:
          type: ["number", "null"]
        profileCompleteness:
          type: integer
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        recentReviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"