# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  doulaId: number;
  /** @nullable */
  doulaName?: string | null;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  clientPhone?: string | null;
  serviceType: string;
  /** @nullable */
  dueDate?: string | null;
  status: BookingStatus;
  message: string;
  /** @nullable */
  doulaResponse?: string | null;
  createdAt: string;
}