# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  doulaId: number;
  clientName: string;
  clientEmail: string;
  clientPhone?: string;
  serviceType: string;
  dueDate?: string;
  message: string;
}