# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateStatus } from "./bookingUpdateStatus";

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  doulaResponse?: string;
}