# lib/api-zod/src/generated/types/doulaDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";
import type { Review } from "./review";

export interface DoulaDashboard {
  doulaId: number;
  pendingBookings: number;
  acceptedBookings: number;
  completedBookings: number;
  totalReviews: number;
  /** @nullable */
  averageRating: number | null;
  profileCompleteness: number;
  recentBookings: Booking[];
  recentReviews: Review[];
}