# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookYourDoula API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  doulaId: number;
  clientName: string;
  rating: number;
  /** @nullable */
  title?: string | null;
  body: string;
  serviceType: string;
  /** @nullable */
  birthYear?: number | null;
  verified?: boolean;
  createdAt: string;
}